requirements.txt
streamlit
python-dotenv
firecrawl-py